feat: drivers/cpufreq/exynos-acme.c: Finalizing Overclock changes + test for GPU overclock
Credits: 	@DaemonMCR (Thomas Turner) for the GPU OC/UC
	@physwizz, Chatur, and Carlos Burero for the CPU OC/UC

This commit should finally allow the Exynos CPU (and GPU?) to finally
go to higher frequency than Samsung Allowed.

However, this may introduce some heat problems which i think
the other components on the kernel should try taking care of it
(throtlling) to reduce damage.

As for Thomas's (@DaemonMCR)'s GPU OC/UC, this is the first one
from MizProject to utilize this component. If this works, it's all
great, if not, then bummer.

Signed-off-by: Mizumo-prjkt <bluepearl394@gmail.com>